move State to state.rs
